Infrastructure minimal refactor
